Move env vars to custom agents
